+++ vue/sbc-common-components/src/services/token.services.ts
@@ -1,137 +1,15 @@
-import Keycloak, { KeycloakInitOptions, KeycloakInstance } from 'keycloak-js'
-import ConfigHelper from '../util/config-helper'
-import { SessionStorageKeys } from '../util/constants'
-import { getModule } from 'vuex-module-decorators'
-import AuthModule from '../store/modules/auth'
 import { Store } from 'vuex'
+import KeycloakServices from './keycloak.services'
+
+// TODO: Remove the whole file once the other teams changed from token-service to keycloak-service
 
 class TokenServices {
-  private kc: KeycloakInstance | undefined
-  private counter = 0
-  private REFRESH_ATTEMPT_INTERVAL = 10 // in seconds
-  private timerId = 0
-  private store: Store<any> | null = null
-
   async init (store?: Store<any>, isScheduleRefresh: boolean = true) {
-    this.store = store
-    const kcOptions: KeycloakInitOptions = {
-      onLoad: 'check-sso',
-      checkLoginIframe: false,
-      timeSkew: 0,
-      token: ConfigHelper.getFromSession(SessionStorageKeys.KeyCloakToken) || undefined,
-      refreshToken: ConfigHelper.getFromSession(SessionStorageKeys.KeyCloakRefreshToken) || undefined,
-      idToken: ConfigHelper.getFromSession(SessionStorageKeys.KeyCloakIdToken) || undefined
-    }
-
-    return new Promise((resolve, reject) => {
-      this.kc = Keycloak(ConfigHelper.getKeycloakConfigUrl())
-      ConfigHelper.addToSession(SessionStorageKeys.SessionSynced, false)
-      this.kc.init(kcOptions)
-        .success(authenticated => {
-          console.info('[TokenServices] is User Authenticated?: Syncing ' + authenticated)
-          if (this.kc && authenticated) {
-            ConfigHelper.addToSession(SessionStorageKeys.SessionSynced, true)
-            this.syncSessionStorage()
-            if (isScheduleRefresh) {
-              this.scheduleRefreshTimer()
-            }
-            resolve(this.kc.token)
-          } else {
-            // If not authenticated that means token is invalid
-            // Clear out session storage and go to auth home (TODO: Perhaps make this a propery parent apps could pass in?)
-            this.clearSession()
-            reject(new Error('NOT_AUTHENTICATED'))
-          }
-        })
-        .error(error => {
-          reject(new Error('Could not Initialize KC' + error))
-        })
-    })
+    return KeycloakServices.initializeToken(store, isScheduleRefresh)
   }
 
   scheduleRefreshTimer (refreshEarlyTime = 0) {
-    let refreshEarlyTimeinMilliseconds = Math.max(this.REFRESH_ATTEMPT_INTERVAL, refreshEarlyTime) * 1000
-    this.scheduleRefreshToken(refreshEarlyTimeinMilliseconds)
-  }
-
-  refreshToken () {
-    console.log('[TokenServices] One time Token Refreshing ')
-    return new Promise((resolve, reject) => {
-      if (this.kc) {
-        this.kc.updateToken(-1)
-          .success(refreshed => {
-            if (refreshed) {
-              console.log('[TokenServices] One time Token Refreshed ')
-              this.syncSessionStorage()
-              resolve()
-            }
-          })
-          .error(() => {
-            reject(new Error('Could not refresh Token'))
-          })
-      } else {
-        reject(new Error('Could not refresh Token:No Kc Instance'))
-      }
-    })
-  }
-
-  private scheduleRefreshToken (refreshEarlyTimeinMilliseconds: number) {
-    let refreshTokenExpiresIn = -1
-    // check if refresh token is still valid . Or else clear all timers and throw errors
-    if (this.kc && this.kc.timeSkew !== undefined && this.kc.refreshTokenParsed) {
-      refreshTokenExpiresIn = this.kc.refreshTokenParsed['exp']! - Math.ceil(new Date().getTime() / 1000) + this.kc.timeSkew
-    }
-    if (refreshTokenExpiresIn < 0) {
-      throw new Error('Refresh Token Expired. No more token refreshes')
-    }
-    let expiresIn = -1
-    if (this.kc && this.kc.tokenParsed && this.kc.tokenParsed['exp'] && this.kc.timeSkew !== undefined) {
-      expiresIn = this.kc.tokenParsed['exp'] - Math.ceil(new Date().getTime() / 1000) + this.kc.timeSkew
-    }
-    if (expiresIn < 0) {
-      throw new Error('Refresh Token Expired. No more token refreshes')
-    }
-    let refreshInMilliSeconds = (expiresIn * 1000) - refreshEarlyTimeinMilliseconds // in milliseconds
-    console.info('[TokenServices] Token Refresh Scheduled in %s Seconds', (refreshInMilliSeconds / 1000))
-    this.timerId = setTimeout(() => {
-      console.log('[TokenServices] Refreshing Token Attempt: %s ', ++this.counter)
-      this.kc!.updateToken(-1)
-        .success(refreshed => {
-          if (refreshed) {
-            console.log('Token successfully refreshed')
-            this.syncSessionStorage()
-            this.scheduleRefreshToken(refreshEarlyTimeinMilliseconds)
-          }
-        })
-        .error(() => {
-          clearTimeout(this.timerId)
-        })
-    }, refreshInMilliSeconds)
-  }
-
-  private syncSessionStorage () {
-    if (this.kc) {
-      if (this.kc.token) {
-        ConfigHelper.addToSession(SessionStorageKeys.KeyCloakToken, this.kc.token)
-      }
-      if (this.kc.refreshToken) {
-        ConfigHelper.addToSession(SessionStorageKeys.KeyCloakRefreshToken, this.kc.refreshToken)
-      }
-      if (this.kc.idToken) {
-        ConfigHelper.addToSession(SessionStorageKeys.KeyCloakIdToken, this.kc.idToken)
-      }
-    }
-  }
-
-  private async clearSession () {
-    if (this.store) {
-      const authModule = getModule(AuthModule, this.store)
-      authModule.clearSession()
-    } else {
-      ConfigHelper.removeFromSession(SessionStorageKeys.KeyCloakToken)
-      ConfigHelper.removeFromSession(SessionStorageKeys.KeyCloakIdToken)
-      ConfigHelper.removeFromSession(SessionStorageKeys.KeyCloakRefreshToken)
-    }
+    KeycloakServices.scheduleRefreshTimer(refreshEarlyTime)
   }
 }
